refactor: Refactor to use ResponseDto(#30)
