--- src/main/java/com/sogonsogon/neighclova/dto/response/auth/CheckPasswordResponseDto.java
@@ -1,36 +1,0 @@
-package com.sogonsogon.neighclova.dto.response.auth;
-
-import com.sogonsogon.neighclova.common.ResponseCode;
-import com.sogonsogon.neighclova.common.ResponseMessage;
-import com.sogonsogon.neighclova.dto.response.ResponseDto;
-import lombok.Getter;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-
-@Getter
-public class CheckPasswordResponseDto extends ResponseDto {
-
-    private CheckPasswordResponseDto() {
-        super(ResponseCode.SUCCESS, ResponseMessage.SUCCESS);
-    }
-
-    public static ResponseEntity<CheckPasswordResponseDto> success() {
-        CheckPasswordResponseDto result = new CheckPasswordResponseDto();
-        return ResponseEntity.status(HttpStatus.OK).body(result);
-    }
-
-    public static ResponseEntity<ResponseDto> notExistUser() {
-        ResponseDto result = new ResponseDto(ResponseCode.NOT_EXISTED_USER, ResponseMessage.NOT_EXISTED_USER);
-        return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(result);
-    }
-
-    public static ResponseEntity<ResponseDto> noPermission() {
-        ResponseDto result = new ResponseDto(ResponseCode.NO_PERMISSION, ResponseMessage.NO_PERMISSION);
-        return ResponseEntity.status(HttpStatus.FORBIDDEN).body(result);
-    }
-
-    public static ResponseEntity<ResponseDto> noAuthentication() {
-        ResponseDto result = new ResponseDto(ResponseCode.AUTHORIZATION_FAIL, ResponseMessage.AUTHORIZATION_FAIL);
-        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(result);
-    }
-}
