--- src/main/java/com/sogonsogon/neighclova/dto/response/auth/PatchPasswordResponseDto.java
@@ -1,31 +1,0 @@
-package com.sogonsogon.neighclova.dto.response.auth;
-
-import com.sogonsogon.neighclova.common.ResponseCode;
-import com.sogonsogon.neighclova.common.ResponseMessage;
-import com.sogonsogon.neighclova.dto.response.ResponseDto;
-import lombok.Getter;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-
-@Getter
-public class PatchPasswordResponseDto extends ResponseDto{
-
-    private PatchPasswordResponseDto() {
-        super(ResponseCode.SUCCESS, ResponseMessage.SUCCESS);
-    }
-
-    public static ResponseEntity<PatchPasswordResponseDto> success() {
-        PatchPasswordResponseDto result = new PatchPasswordResponseDto();
-        return ResponseEntity.status(HttpStatus.OK).body(result);
-    }
-
-    public static ResponseEntity<ResponseDto> notExistUser() {
-        ResponseDto result = new ResponseDto(ResponseCode.NOT_EXISTED_USER, ResponseMessage.NOT_EXISTED_USER);
-        return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(result);
-    }
-
-    public static ResponseEntity<ResponseDto> noAuthentication() {
-        ResponseDto result = new ResponseDto(ResponseCode.AUTHORIZATION_FAIL, ResponseMessage.AUTHORIZATION_FAIL);
-        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(result);
-    }
-}
