+++ src/main/java/com/sogonsogon/neighclova/service/AuthService.java
@@ -1,6 +1,7 @@
 package com.sogonsogon.neighclova.service;
 
 import com.sogonsogon.neighclova.dto.request.auth.*;
+import com.sogonsogon.neighclova.dto.response.ResponseDto;
 import com.sogonsogon.neighclova.dto.response.auth.*;
 import org.springframework.http.ResponseEntity;
 
@@ -17,11 +18,11 @@ public interface AuthService {
 
     ResponseEntity<? super TokenResponseDto> reissue(String refreshToken);
 
-    ResponseEntity<? super PatchPasswordResponseDto> patchPassword(PatchPasswordRequestDto dto, String email);
+    ResponseEntity<ResponseDto> patchPassword(PatchPasswordRequestDto dto, String email);
 
     ResponseEntity<? super DeleteUserResponseDto> deleteUser(String email);
 
-    ResponseEntity<? super CheckPasswordResponseDto> checkPassword(CheckPasswordRequestDto dto, String email);
+    ResponseEntity<ResponseDto> checkPassword(CheckPasswordRequestDto dto, String email);
 
-    ResponseEntity<? super CheckSocialResponseDto> checkSocial(String email);
+    ResponseEntity<ResponseDto> checkSocial(String email);
 }
